+++ src/core/request/engines/provider/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.34.1 (2021-03-04)
+
+#### :bug: Bug Fix
+
+* Fixed incorrect path resolving for providers chain with parameters in a URL
+
 ## v3.33.0 (2021-02-26)
 
 #### :rocket: New Feature